requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
PyYAML>=5.4.0